pyproject.toml
[tool.poetry]
name = "drf_auth"
version = "0.1.0"
description = ""
authors = ["MohamadSheikhAlshabab <alshabab.moh@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
django = "^3.1.1"
djangorestframework = "^3.12.1"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
djangorestframework_simplejwt = "^4.4.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
